requirements.txt
Django==1.6.5
South==1.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-geoposition==0.2.1
django-oauth2-provider==0.2.6.1
django-tastypie==0.11.1
django-toolbelt==0.0.1
gunicorn==19.1.0
hashlib==20081119
mimeparse==0.1.3
psycopg2==2.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
shortuuid==0.4.2
six==1.7.3
static3==0.5.1
wsgiref==0.1.2
